# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5Swagger",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local"
        },
        {
            "url": "http://localhost/api/auth",
            "description": "Production"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "Арендаторы"
                ],
                "summary": "Получить всех пользователей",
                "description": "Получить пользователей",
                "operationId": "a09b4a9e85ae460c4259bd91492dfcad",
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "Арендаторы"
                ],
                "summary": "Получить пользователя",
                "description": "Получает пользователя по идентификатору и возвращает его",
                "operationId": "6e12f4621a3f7f30632c7f9ed096904c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "deb4ff7a-c16b-4b9f-98db-d3c4e3cda010"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "tags": [
                    "Арендаторы"
                ],
                "summary": "Создать пользователя",
                "description": "Создает нового пользователя и возвращает его",
                "operationId": "853f5a91dbf8a9e58c62b951eecf4097",
                "parameters": [
                    {
                        "name": "default_bill_id",
                        "in": "path",
                        "description": "Cчет по умолчанию",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "5z7490a8-f20e-32eb-87f4-3630d5999c0b"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "path",
                        "description": "Фамилия арендатора",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Haley"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "path",
                        "description": "Имя арендатора",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Carolyn"
                        }
                    },
                    {
                        "name": "middle_name",
                        "in": "path",
                        "description": "Отчество арендатора",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Berta"
                        }
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Статус аккаунта",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "active"
                        }
                    },
                    {
                        "name": "passport_series",
                        "in": "path",
                        "description": "Серия паспорта",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "21 52"
                        }
                    },
                    {
                        "name": "passport_number",
                        "in": "path",
                        "description": "Номер паспорта",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "026907"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "Номер телефона",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "7525301782"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserCreate"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response401"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response403"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Response404"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "uuid",
                        "example": "af42801a-70bb-4966-87d6-d53ead3015b5"
                    },
                    "default_bill_id": {
                        "type": "uuid",
                        "example": "5z7490a8-f20e-32eb-87f4-3630d5999c0b"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Haley"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Carolyn"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "Berta"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "passport_series": {
                        "type": "string",
                        "example": "21 52"
                    },
                    "passport_number": {
                        "type": "string",
                        "example": "026907"
                    },
                    "phone": {
                        "type": "integer",
                        "example": "7525301782"
                    }
                },
                "type": "object"
            },
            "UserCreate": {
                "title": "UserCreate",
                "properties": {
                    "default_bill_id": {
                        "type": "uuid",
                        "example": "5z7490a8-f20e-32eb-87f4-3630d5999c0b"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Haley"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Carolyn"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "Berta"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "passport_series": {
                        "type": "string",
                        "example": "21 52"
                    },
                    "passport_number": {
                        "type": "string",
                        "example": "026907"
                    },
                    "phone": {
                        "type": "integer",
                        "example": "7525301782"
                    }
                },
                "type": "object"
            },
...
            "Response401": {
                "title": "Response401",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
            "Response403": {
                "title": "Response403",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden"
                    }
                },
                "type": "object"
            },
            "Response404": {
                "title": "Response404",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}